07_streamingchatbot-with-chainlite/pyproject.toml
[project]
name = "07-streamingchatbot-with-chainlite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.6.2",
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
]
